Add files via upload
In this version of the project, the Network class is introduced for creating a network of nodes to connect with one another before being able to then display the shortest pathway between two nodes in the network. All the functions that were built to contribute in finding the shortest path between two nodes have been copied from the Program.cs file into the Network class. Most of those functions are now private, apart from the getPath function. A further change that must be made in the Network class is to rename everywhere where it says nodes to this.nodes now that the nodes list is inside the Network class.